+++ src/Check/TinyPngCheck.php
@@ -14,10 +14,10 @@ declare(strict_types=1);
 namespace Ekino\TinyPngSonataMediaBundle\Check;
 
 use Ekino\TinyPngSonataMediaBundle\Client\ClientInterface;
-use ZendDiagnostics\Check\CheckInterface;
-use ZendDiagnostics\Result\Failure;
-use ZendDiagnostics\Result\ResultInterface;
-use ZendDiagnostics\Result\Success;
+use Laminas\Diagnostics\Check\CheckInterface;
+use Laminas\Diagnostics\Result\Failure;
+use Laminas\Diagnostics\Result\ResultInterface;
+use Laminas\Diagnostics\Result\Success;
 
 /**
  * Class TinyPngCheck